+++ activity_cli/cmd_edit.h
@@ -11,7 +11,7 @@ public:
     int runCommand(const std::vector<std::string>& args) override;
     std::string syntax() const override;
     std::string description() const override;
-    size_t minNumberOfArgs() const override;
-    size_t maxNumberOfArgs() const override;
+    int minNumberOfArgs() const override;
+    int maxNumberOfArgs() const override;
 };
