+++ activity_cli/cmd_get.h
@@ -11,6 +11,6 @@ public:
     int runCommand(const std::vector<std::string>& args) override;
     std::string syntax() const override;
     std::string description() const override;
-    size_t minNumberOfArgs() const override;
-    size_t maxNumberOfArgs() const override;
+    int minNumberOfArgs() const override;
+    int maxNumberOfArgs() const override;
 };
